Ajustado polimorfismo do player e enemy
