+++ README.md
@@ -4,3 +4,6 @@
 
 - Linha Direta The Game é um rogue-like estilo top-down ambientado no Rio de Janeiro, no qual você deve derrotar o chefe do tráfico, passando por uma série de ambientes selecionados aleatoriamente até o chefe final.
 A cada sala, o jogador recebe alguns itens, podendo equipá-los ou não, assim melhorando seus status para finalizar o jogo mais rápido.
+
+## Dependências 
+- pip install pygame
